Reorganizes file structure of code and adds script equivalencies of notebook code.
For users wanting to replicate results, it is best if they only needed to run python files then notebooks. I copied all of the code from the notebooks into the Python scripts, and changed the file structure and code to keep the root of the project clean. download_datasets.ipynb needs to be read if one is to reproduce the project, so I kept that in the root.