Refactors AdaFace.
My AdaFace modification did not properly keep the PyTorch structure from the original project. This commit primarily re-writes CustomAdaFace. In addition, I re-organized the files from AdaFace I needed, and commented out the ViTPose deps from requirements.txt as we may be able to get away with it with the new way we use the model and it may have been causing a circular import error in Pytorch Lightning due to so many deps.