Removes warning for adaface imports.
Due to the location of the adaface code relative to the code using it, Python cannot natively find the files. I already have code blocks set up to add the library to the project path, which is used so Python knows where the files are located, but this still gives a semantics warning in Pylance. To fix this, I am adding the lib folder as an extra path in settings.json so this warning will go away.