Improves CustomYOLO
Now that I changed the code for the backbone, now I need to go back and change YOLO's input as well. I moved to a more gradual reduction method, which is similar to what the original paper does. Additionally, I edited the forward function similar to in AdaFace. I also added a code block for processing the results once it is done if it is in inference mode.