Updates Download models script.
As we are changing models in HF, we need to change the install script. I updates the script to download the YOLOv11 models instead of v8. In addition, I also decided to make the code a bit more explicit to readers, by ignoring the incorrect `request` import error and adding typing to the method as well as putting the args in the doc comments.