Edits CustomYOLO for v11 and spelling.
In addition to adjusting the path to the original model, the gradual adapter needed to be re-considered  for the structural differences between the models. Specifically, v11 has an additional layer that is an early feature processing block that is optimal to be removed due to the backbone. In addition, this commit also runs the spell checker on the text cells in the notebook and edits accordingly.